Skip steps if django_comments does not exist.
Required for migrations to work properly.

Was previously created by the django.contrib.comments module which is no
longer used.

Patch provided by Russell Sim.

+++ karaage/people/migrations/0018_add_fields_person.py
@@ -2,7 +2,7 @@
 import datetime
 from south.db import db
 from south.v2 import SchemaMigration
-from django.db import models
+from django.db import models, connection
 
 
 class Migration(SchemaMigration):
@@ -24,12 +24,17 @@ class Migration(SchemaMigration):
         db.add_column('django_admin_log', 'person',
                       self.gf('django.db.models.fields.related.ForeignKey')(to=orm['people.person'], null=True),
                       keep_default=False)
-        db.add_column('django_comments', 'person',
-                      self.gf('django.db.models.fields.related.ForeignKey')(to=orm['people.person'], null=True),
-                      keep_default=False)
-        db.add_column('django_comment_flags', 'person',
-                      self.gf('django.db.models.fields.related.ForeignKey')(to=orm['people.person'], null=True),
-                      keep_default=False)
+
+        # If the django comments table exists then add the extra missing fields
+        cursor = connection.cursor()
+        if 'django_comments' in connection.introspection.get_table_list(cursor):
+            db.add_column('django_comments', 'person',
+                          self.gf('django.db.models.fields.related.ForeignKey')(to=orm['people.person'], null=True),
+                          keep_default=False)
+        if 'django_comment_flags' in connection.introspection.get_table_list(cursor):
+            db.add_column('django_comment_flags', 'person',
+                          self.gf('django.db.models.fields.related.ForeignKey')(to=orm['people.person'], null=True),
+                          keep_default=False)
 
         # Adding field 'Person.password'
         db.add_column('person', 'password',
@@ -77,8 +82,12 @@ class Migration(SchemaMigration):
 
     def backwards(self, orm):
         db.delete_column('django_admin_log', 'person_id')
-        db.delete_column('django_comments', 'person_id')
-        db.delete_column('django_comment_flags', 'person_id')
+
+        cursor = connection.cursor()
+        if 'django_comments' in connection.introspection.get_table_list(cursor):
+            db.delete_column('django_comments', 'person_id')
+        if 'django_comment_flags' in connection.introspection.get_table_list(cursor):
+            db.delete_column('django_comment_flags', 'person_id')
 
         # Deleting field 'Person.password'
         db.delete_column('person', 'password')
